# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cognexa API
  version: 1.0.0
  description: |
    Canonical HTTP contract for the supported Cognexa TypeScript API.
    Unversioned root aliases are deprecated and sunset on 2026-12-31.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.cognexa.app/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /courses:
    get:
      tags: [Courses]
      security: []
      summary: List public courses or the authenticated user's course views
      parameters:
        - name: filter
          in: query
          schema: { $ref: '#/components/schemas/CourseStatus' }
      responses:
        '200':
          description: Course views; private enrollment data is removed for anonymous callers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Course' }
    post:
      tags: [Courses]
      summary: Create a course and enroll the creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseName]
              properties:
                courseName: { type: string, minLength: 1 }
                description: { type: string }
                image: { type: string, format: uri }
      responses:
        '201':
          description: Updated course collection
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /courses/{courseId}/lectures:
    get:
      tags: [Curriculum]
      summary: List course video metadata for an enrolled learner or course manager
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200': { description: Video metadata }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /instructor/courses:
    post:
      tags: [Instructor]
      summary: Create a persisted course draft owned by the current instructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 1, maxLength: 160, default: Untitled course }
      responses:
        '201':
          description: Created draft workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseWorkspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /deadlines:
    get:
      tags: [Courses]
      summary: List deadlines for the current user's courses
      responses:
        '200': { description: Deadlines }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token held in browser memory only.
    refreshCookie:
      type: apiKey
      in: cookie
      name: cognexa_refresh
      description: Opaque rotating HttpOnly session credential.
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
...
  responses:
...
    Unauthorized:
      description: Authentication required or session expired
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated but not authorized
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ValidationError:
      description: Request validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
        requestId: { type: string }
...
    Course:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        image: { type: string, format: uri }
        backgroundColor: { type: string }
        status: { $ref: '#/components/schemas/CourseStatus' }
        enrolled: { type: boolean }
        privilege: { type: string, enum: [student, instructor, admin] }
        createdBy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            username: { type: string }
            photo: { type: string, format: uri }
    CourseStatus:
      type: string
      enum: [draft, review, published, archived]
    UploadedAsset:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url: { type: string, format: uri }
        publicId: { type: string }
        resourceType: { type: string, enum: [image, video, raw] }
        format: { type: string }
        bytes: { type: integer, minimum: 0 }
        originalName: { type: string }
        thumbnailUrl: { type: string, format: uri }
    CourseLesson:
      type: object
      additionalProperties: false
      required: [title, type, order, isPreview]
      properties:
        id: { type: string }
        title: { type: string, minLength: 1, maxLength: 160 }
        type:
          type: string
          enum: [video, pdf, markdown, rich_text, external_url, youtube, live_session, file]
        order: { type: integer, minimum: 0 }
        description: { type: string, maxLength: 2000 }
        url: { type: string, format: uri }
        content: { type: string, maxLength: 100000 }
        durationMinutes: { type: integer, minimum: 0 }
        isPreview: { type: boolean }
        asset: { $ref: '#/components/schemas/UploadedAsset' }
    CourseModule:
      type: object
      additionalProperties: false
      required: [title, order, moduleItems]
      properties:
        id: { type: string }
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        order: { type: integer, minimum: 0 }
        moduleItems:
          type: array
          maxItems: 250
          items: { $ref: '#/components/schemas/CourseLesson' }
    BuilderQuestion:
      type: object
      additionalProperties: false
      required: [prompt, type, options, correctAnswers, points]
      properties:
        id: { type: string }
        prompt: { type: string, minLength: 1, maxLength: 10000 }
        type: { type: string, enum: [mcq, multiple_select, true_false, fill_blank] }
        options: { type: array, maxItems: 20, items: { type: string } }
        correctAnswers: { type: array, maxItems: 20, items: { type: string } }
        explanation: { type: string, maxLength: 10000 }
        points: { type: integer, minimum: 1, maximum: 1000 }
        pool: { type: string, maxLength: 120 }
    BuilderAssessment:
      type: object
      additionalProperties: false
      required:
        [
          kind,
          title,
          order,
          visibility,
          questions,
          randomizeQuestions,
          randomizeAnswers,
          passingScore,
          attachments,
          rubric,
          submissionLimit,
        ]
      properties:
        id: { type: string }
        kind: { type: string, enum: [quiz, assignment] }
        title: { type: string, minLength: 1, maxLength: 160 }
        instructions: { type: string, maxLength: 100000 }
        order: { type: integer, minimum: 0 }
        visibility: { type: string, enum: [draft, published] }
        questions: { type: array, maxItems: 500, items: { $ref: '#/components/schemas/BuilderQuestion' } }
        randomizeQuestions: { type: boolean }
        randomizeAnswers: { type: boolean }
        passingScore: { type: number, minimum: 0, maximum: 100 }
        timeLimitMinutes: { type: integer, minimum: 1, maximum: 1440 }
        questionPoolSize: { type: integer, minimum: 1, maximum: 1000 }
        dueDate: { type: string, format: date-time }
        attachments: { type: array, maxItems: 20, items: { $ref: '#/components/schemas/UploadedAsset' } }
        rubric:
          type: array
          maxItems: 50
          items:
            type: object
            required: [title, points]
            properties:
              id: { type: string }
              title: { type: string }
              description: { type: string }
              points: { type: number, minimum: 0 }
        submissionLimit: { type: integer, minimum: 1, maximum: 100 }
    CourseWorkspaceInput:
      type: object
      additionalProperties: false
      required: [name, tags, language, level, prerequisites, pricing, seo, modules, assessments]
      properties:
        draftVersion: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 160 }
        subtitle: { type: string, maxLength: 240 }
        description: { type: string, maxLength: 20000 }
        image: { type: string, format: uri }
        thumbnail: { $ref: '#/components/schemas/UploadedAsset' }
        banner: { $ref: '#/components/schemas/UploadedAsset' }
        category: { type: string, maxLength: 120 }
        tags: { type: array, maxItems: 30, items: { type: string } }
        language: { type: string }
        level: { type: string, enum: [beginner, intermediate, advanced, all-levels] }
        durationMinutes: { type: integer, minimum: 0 }
        prerequisites: { type: array, maxItems: 50, items: { type: string } }
        pricing:
          type: object
          additionalProperties: false
          required: [model, amount, currency]
          properties:
            model: { type: string, enum: [free, paid] }
            amount: { type: number, minimum: 0 }
            currency: { type: string, minLength: 3, maxLength: 3 }
        seo:
          type: object
          additionalProperties: false
          properties:
            title: { type: string, maxLength: 70 }
            description: { type: string, maxLength: 170 }
            slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        modules: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/CourseModule' } }
        assessments: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/BuilderAssessment' } }
    CourseWorkspace:
      allOf:
        - $ref: '#/components/schemas/CourseWorkspaceInput'
        - type: object
          required: [id, status, draftVersion]
          properties:
            id: { type: string }
            status: { $ref: '#/components/schemas/CourseStatus' }
            reviewNotes: { type: string }
            reviewSubmittedAt: { type: string, format: date-time }
            publishedAt: { type: string, format: date-time }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...